django/pyproject.toml
[tool.poetry]
name = "rag-n-chat"
version = "0.1.0"
description = "Rag And Chat API"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiohappyeyeballs = "2.4.3"
aiohttp = "3.10.10"
aiosignal = "1.3.1"
attrs = "24.2.0"
certifi = "2024.8.30"
charset-normalizer = "3.4.0"
et-xmlfile = "1.1.0"
frozenlist = "1.5.0"
idna = "3.10"
multidict = "6.1.0"
numpy = "2.1.2"
openai = "0.28.0"
openpyxl = "3.1.5"
pandas = "2.2.3"
pandas-stubs = "2.2.3.241009"
pinecone = "5.3.1"
pinecone-plugin-inference = "1.1.0"
pinecone-plugin-interface = "0.0.7"
propcache = "0.2.0"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
pytz = "2024.2"
requests = "2.32.3"
six = "1.16.0"
tqdm = "4.66.5"
types-pytz = "2024.2.0.20241003"
typing-extensions = "4.12.2"
tzdata = "2024.2"
urllib3 = "2.2.3"
yarl = "1.16.0"
django = "^5.1.2"
djangorestframework = "^3.15.2"
django-environ = "^0.11.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
